+++ tests/system/test_gbq.py
@@ -314,7 +314,7 @@ class TestReadGBQIntegration(object):
             df,
             DataFrame(
                 {"unix_epoch": ["1970-01-01T00:00:00.000000Z"]},
-                dtype="datetime64[ns]",
+                dtype=pandas.api.types.DatetimeTZDtype(tz="UTC"),
             ),
         )
 
@@ -330,7 +330,7 @@ class TestReadGBQIntegration(object):
             df,
             DataFrame(
                 {"valid_timestamp": ["2004-09-15T05:00:00.000000Z"]},
-                dtype="datetime64[ns]",
+                dtype=pandas.api.types.DatetimeTZDtype(tz="UTC"),
             ),
         )
 
@@ -403,7 +403,11 @@ class TestReadGBQIntegration(object):
             dialect="legacy",
         )
         tm.assert_frame_equal(
-            df, DataFrame({"null_timestamp": [NaT]}, dtype="datetime64[ns]")
+            df,
+            DataFrame(
+                {"null_timestamp": [NaT]},
+                dtype=pandas.api.types.DatetimeTZDtype(tz="UTC"),
+            ),
         )
 
     def test_should_properly_handle_null_datetime(self, project_id):
@@ -589,7 +593,9 @@ class TestReadGBQIntegration(object):
             "title": pandas.Series([], dtype=object),
             "id": pandas.Series([], dtype=np.dtype(int)),
             "is_bot": pandas.Series([], dtype=np.dtype(bool)),
-            "ts": pandas.Series([], dtype="datetime64[ns]"),
+            "ts": pandas.Series(
+                [], dtype=pandas.api.types.DatetimeTZDtype(tz="UTC")
+            ),
         }
         expected_result = DataFrame(
             empty_columns, columns=["title", "id", "is_bot", "ts"]